chore: chanage integer to big int

+++ cmd/tradelogs/migrations/00006_add_column_expiration_date.up.sql
@@ -1,1 +1,1 @@
-ALTER TABLE tradelogs ADD COLUMN IF NOT EXISTS expiration_date INTEGER;
+ALTER TABLE tradelogs ADD COLUMN IF NOT EXISTS expiration_date BIGINT;
